Split up radio and Bot
